SQL schema Complete BL entities according to final SQL schema

+++ RideOnServer/BL/ProductCategory.cs
@@ -2,7 +2,7 @@
 {
     public class ProductCategory
     {
-        public short CategoryId { get; set; }
+        public byte CategoryId { get; set; }
 
         public string CategoryName { get; set; }
     }
